Create Loader.go
Commiting the loader. Instructions on how to use it are in the file header, but the gist is the following: Takes 2 file-paths as arguments, and returns an object of type []training_sample, where training_sample is of type { pointer to pixels []byte, label int }.

Tested on our data, and works!