Added the package where I would like to to gradient desent optimization.
Not much was actually written, because I spent the majority of the time trying various things and trying to figure out
how things work. The biggest 'breakthrough' of this session came in terms on learning about function types, which is
very important for any kind of generic optimization tools.

Specifically, I defined the CostFunction type, and a GradientDesent function which takes such type as an input. By doing that we can construct the CostFunction outside
of this package (in main or even network), and really use the GradientDesent just for optimization.

TODO: Think about how you would access GradientDesent from let's say 'Network' package for optimizations.