I turned back-propagation into a generic one, and it works with a simple test. TrainOffline works as well.
What doesn't work as it should is training with test data -->
there is a bug in FeedForward, since feed forward always results in same output, non-dependent on the training data.
It has to do something with how layer0 neurons output is set. My brain is fried, will take a look tomorrow.